Refactor: refactor and repair card edit logic and component
