+++ src/components/Card/ActionCard/index.tsx
@@ -4,9 +4,9 @@ import { CiCalendarDate } from "react-icons/ci";
 import { HiOutlineDotsVertical } from "react-icons/hi";
 import { SlActionUndo } from "react-icons/sl";
 import { ICard, cardStore } from "../../../store/cardStore";
-import Editable from "../../Editable";
 import BasicCard from "../BasicCard";
-import CardTags from "../CardTags";
+import Editable from "../Editable/Editable";
+import EditableWrapper from "../Editable/EditableWrapper";
 import { IdeaNoteModal } from "../IdeaCard/IdeaToNoteModal";
 import { ActionCardTool } from "./ActionCardTool";
 
@@ -58,9 +58,9 @@ export const ActionCard = ({ card }: CardToolProps) => {
   const isTodo = card.dueDate || card.status === "execute";
 
   return (
-    <BasicCard card={card}>
-      <div className="flex flex-grow items-center justify-between">
-        <div className="flex flex-grow items-center justify-between">
+    <>
+      <BasicCard card={card}>
+        <EditableWrapper>
           {isTodo && (
             <Checkbox
               size="sm"
@@ -70,7 +70,7 @@ export const ActionCard = ({ card }: CardToolProps) => {
               isSelected={isSelect}
               lineThrough
               defaultSelected
-            ></Checkbox>
+            />
           )}
           <Editable
             id={card.id}
@@ -87,28 +87,25 @@ export const ActionCard = ({ card }: CardToolProps) => {
               ref={inputRef}
             />
           </Editable>
-        </div>
-        <ul className="ml-2 flex min-w-unit-24 items-center justify-between">
-          {settingList.map((setting) => (
-            <li id={setting.id} key={setting.label}>
-              <ActionCardTool
-                setting={setting}
-                onOpen={onOpenNote}
-                card={card}
-              />
-            </li>
-          ))}
-        </ul>
-        <IdeaNoteModal
-          card={card}
-          isOpen={isOpenNote}
-          onOpenChange={onOpenChangeNote}
-          onClose={onCloseNote}
-        />
-      </div>
-      <div className="mt-2 flex flex-wrap items-center gap-x-2">
-        <CardTags card={card} />
-      </div>
-    </BasicCard>
+          <ul className="ml-2 flex min-w-unit-24 items-center justify-between">
+            {settingList.map((setting) => (
+              <li id={setting.id} key={setting.label}>
+                <ActionCardTool
+                  setting={setting}
+                  onOpen={onOpenNote}
+                  card={card}
+                />
+              </li>
+            ))}
+          </ul>
+        </EditableWrapper>
+      </BasicCard>
+      <IdeaNoteModal
+        card={card}
+        isOpen={isOpenNote}
+        onOpenChange={onOpenChangeNote}
+        onClose={onCloseNote}
+      />
+    </>
   );
 };
