+++ src/components/Card/Editable/EditableWrapper.tsx
@@ -1,0 +1,9 @@
+function EditableWrapper({ children }: { children: React.ReactNode }) {
+  return (
+    <div className="flex flex-grow items-center justify-between">
+      {children}
+    </div>
+  );
+}
+
+export default EditableWrapper;
