+++ src/components/Card/IdeaCard/index.tsx
@@ -5,9 +5,9 @@ import { HiOutlineDotsVertical } from "react-icons/hi";
 import { PiNoteBlankThin } from "react-icons/pi";
 import { ICard } from "../../../store/cardStore";
 import { uiStore } from "../../../store/uiStore";
-import Editable from "../../Editable";
 import BasicCard from "../BasicCard";
-import CardTags from "../CardTags";
+import Editable from "../Editable/Editable";
+import EditableWrapper from "../Editable/EditableWrapper";
 import IdeaCardTool from "./IdeaCardTool";
 import { IdeaToActionModal } from "./IdeaToActionModal";
 import { IdeaNoteModal } from "./IdeaToNoteModal";
@@ -49,46 +49,45 @@ export const IdeaCard = ({ card }: { card: ICard }) => {
   ];
 
   return (
-    <BasicCard card={card}>
-      <div className="flex flex-grow items-center justify-between">
-        <Editable
-          id={card.id}
-          text={card.content}
-          placeholder="暫無內容..."
-          childRef={inputRef}
-          type="input"
-        >
-          <input
-            className="inline-block w-full bg-transparent tracking-wide outline-none"
-            type="text"
-            name={card.status}
-            defaultValue={card.content}
-            ref={inputRef}
-          />
-        </Editable>
-        <ul className="ml-2 flex min-w-unit-24 items-center justify-between">
-          {settingList.map((setting) => (
-            <li id={setting.id} key={setting.label}>
-              <IdeaCardTool card={card} setting={setting} />
-            </li>
-          ))}
-          <IdeaToActionModal
-            card={card}
-            isOpen={isOpenAction}
-            onOpenChange={onOpenChangeAction}
-            onClose={onCloseAction}
-          />
-          <IdeaNoteModal
-            card={card}
-            isOpen={isOpenNote}
-            onOpenChange={onOpenChangeNote}
-            onClose={onCloseNote}
-          />
-        </ul>
-      </div>
-      <div className="mt-5 flex flex-wrap items-center gap-x-2">
-        <CardTags card={card} />
-      </div>
-    </BasicCard>
+    <>
+      <BasicCard card={card}>
+        <EditableWrapper>
+          <Editable
+            id={card.id}
+            text={card.content}
+            placeholder="暫無內容..."
+            childRef={inputRef}
+            type="input"
+          >
+            <input
+              className="inline-block w-full bg-transparent tracking-wide outline-none"
+              type="text"
+              name={card.status}
+              defaultValue={card.content}
+              ref={inputRef}
+            />
+          </Editable>
+          <ul className="ml-2 flex min-w-unit-24 items-center justify-between">
+            {settingList.map((setting) => (
+              <li id={setting.id} key={setting.label}>
+                <IdeaCardTool card={card} setting={setting} />
+              </li>
+            ))}
+          </ul>
+        </EditableWrapper>
+      </BasicCard>
+      <IdeaToActionModal
+        card={card}
+        isOpen={isOpenAction}
+        onOpenChange={onOpenChangeAction}
+        onClose={onCloseAction}
+      />
+      <IdeaNoteModal
+        card={card}
+        isOpen={isOpenNote}
+        onOpenChange={onOpenChangeNote}
+        onClose={onCloseNote}
+      />
+    </>
   );
 };
