+++ src/components/Card/TodoCard/index.tsx
@@ -4,9 +4,9 @@ import { CiCalendarDate } from "react-icons/ci";
 import { HiOutlineDotsVertical } from "react-icons/hi";
 import { SlActionUndo } from "react-icons/sl";
 import { ICard, cardStore } from "../../../store/cardStore";
-import Editable from "../../Editable";
 import BasicCard from "../BasicCard";
-import CardTags from "../CardTags";
+import Editable from "../Editable/Editable";
+import EditableWrapper from "../Editable/EditableWrapper";
 import { TodoCardTool } from "./TodoCardTool";
 
 const settingList = [
@@ -43,44 +43,39 @@ export const TodoCard = ({ card }: CardToolProps) => {
 
   return (
     <BasicCard card={card}>
-      <div className="flex items-center justify-between">
-        <div className="flex flex-grow">
-          {card.dueDate && (
-            <Checkbox
-              size="sm"
-              radius="sm"
-              name="checkbox"
-              onValueChange={handleComplete}
-              isSelected={isSelect}
-              lineThrough
-              defaultSelected
-            ></Checkbox>
-          )}
-          <Editable
-            id={card.id}
-            text={card.content}
-            placeholder="暫無內容..."
-            childRef={inputRef}
-            type="input"
-          >
-            <input
-              className="inline-block bg-transparent tracking-wide outline-none"
-              type="text"
-              name={card.status}
-              defaultValue={card.content}
-              ref={inputRef}
-            />
-          </Editable>
-        </div>
+      <EditableWrapper>
+        {card.dueDate && (
+          <Checkbox
+            size="sm"
+            radius="sm"
+            name="checkbox"
+            onValueChange={handleComplete}
+            isSelected={isSelect}
+            lineThrough
+            defaultSelected
+          />
+        )}
+        <Editable
+          id={card.id}
+          text={card.content}
+          placeholder="暫無內容..."
+          childRef={inputRef}
+          type="input"
+        >
+          <input
+            className="inline-block w-full bg-transparent tracking-wide outline-none"
+            type="text"
+            name={card.status}
+            defaultValue={card.content}
+            ref={inputRef}
+          />
+        </Editable>
         <div className="ml-2 flex min-w-unit-24 items-center justify-between">
           {settingList.map((setting) => (
             <TodoCardTool key={setting.label} setting={setting} card={card} />
           ))}
         </div>
-      </div>
-      <div className="mt-5 flex flex-wrap items-center gap-x-2">
-        <CardTags card={card} />
-      </div>
+      </EditableWrapper>
     </BasicCard>
   );
 };
